refactor: archi de dossier navigation based
