+++ src/shared/map/MapScreen.tsx
@@ -2,12 +2,12 @@ import React, { useEffect, useState, useContext } from 'react';
 import { View, TouchableOpacity, Linking, ActivityIndicator, StyleSheet, Text } from 'react-native';
 import { WebView } from 'react-native-webview';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
-import Translator from '../../shared/i18n/Translator';
+import Translator from '../i18n/Translator';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 
-import { AppContext } from '../../shared/services/AppCore';
-import style, { tokens } from '../../shared/theme/Theme';
-import { URL } from '../../shared/services/DataService';
+import { AppContext } from '../services/AppCore';
+import style, { tokens } from '../theme/Theme';
+import { URL } from '../services/DataService';
 
 // Import du JSON 
 const locations: Record<string, { lat: number; lng: number }> = require('../../../assets/locations.json');
@@ -41,7 +41,7 @@ export default function MapScreen({ route, navigation }: MapScreenProps) {
         if (typeof loc === 'object' && loc.lat) {
             setLat(loc.lat);
             setLng(loc.lon || loc.lng || null);
-        } 
+        }
         else if (typeof loc === 'string') {
             let house = loc.split('/')[0];
             if (locations.hasOwnProperty(house)) {
@@ -61,8 +61,8 @@ export default function MapScreen({ route, navigation }: MapScreenProps) {
     useEffect(() => {
         if (lat !== null && lng !== null) {
             navigation.setOptions({
-                headerTransparent: true, 
-                title: title || Translator.get('MAP'), 
+                headerTransparent: true,
+                title: title || Translator.get('MAP'),
                 headerRight: () => (
                     <TouchableOpacity
                         onPress={onPressExternalMap}
@@ -140,11 +140,11 @@ export default function MapScreen({ route, navigation }: MapScreenProps) {
 
     return (
         <View style={{ flex: 1, backgroundColor: theme.courseBackground }}>
-            
-            <View style={{ 
-                height: (insets.top || 0) + 65, 
-                backgroundColor: theme.cardBackground, 
-                borderBottomWidth: 1, 
+
+            <View style={{
+                height: (insets.top || 0) + 65,
+                backgroundColor: theme.cardBackground,
+                borderBottomWidth: 1,
                 borderBottomColor: theme.border,
             }} />
